script/requirements.txt
nats-py==2.11.0
curl_cffi==0.13.0
ujson==5.11.0
lxml==6.0.1